# openapi/engine.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeaveKit Engine REST API",
    "version": "0.5.0",
    "description": "The HTTP API exposed by `@weave-kit/engine`. Object routes are generic — `{name}` is an object from your `schema.json`. Every error response uses the same body: `{ \"error\": { \"code\", \"message\", \"params?\" } }`. Authentication is a bearer API key from `auth.source`, and RBAC is enforced in the data-access layer."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key present in `auth.source` (or resolved by an `AuthResolver`)."
      }
    },
    "schemas": {
      "ObjectRecord": {
        "type": "object",
        "additionalProperties": true,
        "description": "A record shaped by the object's `schema.json`."
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable error code, e.g. `rbac.denied.read`."
              },
              "message": {
                "type": "string"
              },
              "params": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "responses": {
      "Error": {
        "description": "Uniform error body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/objects/{name}/{id}/workflow": {
      "get": {
        "tags": [
          "Workflow"
        ],
        "operationId": "getWorkflow",
        "summary": "Workflow status for a record",
        "description": "Returns the record's workflow instance state, current node, the actions available to the caller, and the caller's open workitems. Requires the object to declare `objects/<name>/workflow.json` (otherwise `404`).",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Object name from `objects/<name>/schema.json`."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record id — the `record_key` encoding of the (possibly composite) primary key; returned as `weave_id`."
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "state",
                    "actions"
                  ],
                  "properties": {
                    "state": {
                      "type": "string",
                      "enum": [
                        "draft",
                        "running",
                        "finished",
                        "canceled"
                      ]
                    },
                    "node": {
                      "type": "string"
                    },
                    "approval": {
                      "type": "string",
                      "nullable": true
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "workitems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "nodeId": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Workflow"
        ],
        "operationId": "overrideWorkflow",
        "summary": "Admin override (jump to a node or terminate)",
        "description": "Admin-only (`admin` role). Body `{ node }` jumps the record to that node; `{ state: 'canceled' | 'finished' }` terminates it.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Object name from `objects/<name>/schema.json`."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record id — the `record_key` encoding of the (possibly composite) primary key; returned as `weave_id`."
          }
        ],
        "responses": {
          "200": {
            "description": "Override result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "state": {
                      "type": "string"
                    },
                    "nodeId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/objects/{name}/{id}/workflow/{action}": {
      "post": {
        "tags": [
          "Workflow"
        ],
        "operationId": "runWorkflowAction",
        "summary": "Run a workflow action",
        "description": "Runs `submit` | `approve` | `reject` | `withdraw` | `cancel` | `forward`. RBAC-scoped, workitem-gated, atomic and audited. `forward` takes `{ to: { userId } }`; `approve`/`reject`/`withdraw` take `{ comment }`.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Object name from `objects/<name>/schema.json`."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record id — the `record_key` encoding of the (possibly composite) primary key; returned as `weave_id`."
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Declared workflow transition action name."
          }
        ],
        "responses": {
          "200": {
            "description": "The updated record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectRecord"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  }
}